+++ Assets/Scripts/Interface/IDamagable.cs
@@ -6,7 +6,7 @@ public interface IDamagable
 {
     // 攻撃を受け、ダメージをくらう処理のインターフェース
 
-    void GetAttack(int dmg, Vector2 damagedPosi);
+    void GetAttack(float dmg, Vector2 damagedPosi);
 
     void TakeDamage(int dmg, Vector2 damagedPosi);
 }
